requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langchain==0.3.0
langchain-openai==0.2.0
tavily-python==0.5.0
python-dotenv==1.0.1
pydantic==2.9.0
openai==1.51.0
httpx==0.27.0
python-multipart==0.0.12